Bugfix: Moving While in Combat
While in combat it was possible to move around the map and start other encounters, all with the same health, then fight them all at once.

Changes: Added a variable to FrmMap that indicated whether or not FrmArena was opened. If it was, then the player could not move. Once the window was closed the player could move again.